refactor: move global concurrent settings to a separate class
Close #137
